third_party/blink/web_tests/fragmentation/single-cells-multiple-tables-no-repeating-thead-expected.txt

crbug.com/621258: Don't let repeating headers overlap on pages that contain two tables.

Col 1

Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Col 2
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
PASS
